requirements.txt
# Core Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
mysqlclient==2.2.0  # MySQL connector

# Multi-tenancy
django-tenants==3.5.0
psycopg2-binary==2.9.9  # Requerido por django-tenants aunque usemos MySQL

# Authentication & API
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5  # API documentation

# Development & Debugging
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0  # Test data factories
faker==20.1.0  # Generate fake data for tests

# Environment Variables
python-decouple==3.8

# Excel Generation
xlsxwriter==3.1.9
pandas==2.1.3

# Template Engine
jinja2==3.1.2

# Performance y Caching
redis==5.0.1
django-redis==5.4.0
django-cachalot==2.6.1  # ORM caching

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Monitoring
django-debug-toolbar==4.2.0
sentry-sdk==1.38.0

### Storage y Media
# File Storage
django-storages==1.14.2
boto3==1.34.0  # AWS S3
Pillow==10.1.0  # Image processing

# File Validation
python-magic==0.4.27
django-cleanup==8.0.0  # Auto cleanup files